ml_server/requirements.txt
pyserial>=3.5
websockets>=12.0
numpy>=1.24
